backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Pydantic and Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# File Processing
python-magic>=0.4.27
Pillow>=10.0.0
pdf2image>=1.16.0

# OCR Dependencies
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python>=4.8.0

# LLM Dependencies
# Google Gemini
google-generativeai>=0.3.0

# OpenAI (optional)
openai>=1.3.0

# Anthropic (optional)
anthropic>=0.7.0

# Hugging Face (optional)
transformers>=4.35.0
torch>=2.0.0

# Utilities
python-dateutil>=2.8.0
jsonschema>=4.17.0
python-dotenv>=1.0.0

# Logging and Monitoring
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0 